+++ MiamIOSFramework/MiamIOSFramework/ui/Price/PriceView.swift
@@ -42,8 +42,7 @@ struct PriceView: View {
         } else {
         VStack{
             HStack(alignment: .top, spacing: 2){
-                Text(viewModel.currentState.integerPart+"," ).font(.system(size: 14,weight: .bold))
-                Text(viewModel.currentState.decimalPart + "" + MiamText.sharedInstance.currency).font(.system(size: 14,weight: .bold))
+                Text(String(viewModel.price?.price ?? 0.0)).font(.system(size: 14,weight: .bold));  Text(MiamText.sharedInstance.currency).font(.system(size: 14,weight: .bold))
             }
             if(!isTotalPrice){
                 Text(MiamText.sharedInstance.preGuests).foregroundColor(Color.miamColor(.grey)).font(.system(size: 12))