+++ androidSDK/src/main/java/com/miam/kmm_miam_sdk/android/theme/colors.kt
@@ -3,7 +3,7 @@ package com.miam.kmm_miam_sdk.android.theme
 import androidx.compose.ui.graphics.Color
 
 object Colors {
-    var primary: Color = Color(0xff037E92)
+    var primary: Color = Color(0xff007D8F)
     var secondary: Color = Color(0xFFE61845)
     var ternary: Color = Color(0xFF209B8F)
     var success: Color = Color(0xFF44D6B3)
@@ -11,6 +11,7 @@ object Colors {
     var warning: Color = Color(0xFFFFDAA3)
     var danger: Color = Color(0xffF47F7A)
     var grey: Color = Color(0xFF676767)
+    var backgroundGrey: Color = Color(0xFFF5F4F2)
     var lightgrey: Color = Color(0xFFEDEDED)
     var white: Color = Color(0xFFFAFCFE)
     var unpureWhite: Color = Color(0xFFfefefe)