+++ androidSDK/src/main/java/com/miam/kmm_miam_sdk/android/ui/components/likeButton/LikeButton.kt
@@ -29,8 +29,7 @@ fun LikeButton(
         }
     ) {
         Surface(
-            shape = CircleShape,
-            elevation = 8.dp
+            shape = CircleShape
         ) {
             Box(
                 Modifier