+++ androidSDK/src/main/java/com/miam/kmm_miam_sdk/android/ui/components/price/Price.kt
@@ -1,6 +1,10 @@
 package com.miam.kmm_miam_sdk.android.ui.components.price
 
-import androidx.compose.animation.core.*
+import androidx.compose.animation.core.FastOutLinearInEasing
+import androidx.compose.animation.core.animateFloat
+import androidx.compose.animation.core.infiniteRepeatable
+import androidx.compose.animation.core.rememberInfiniteTransition
+import androidx.compose.animation.core.tween
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -19,8 +23,6 @@ import com.miam.kmmMiamCore.component.pricing.PricingViewModel
 import com.miam.kmm_miam_sdk.android.theme.Typography.bodySmall
 import com.miam.kmm_miam_sdk.android.theme.Typography.subtitleBold
 import com.miam.kmm_miam_sdk.android.ui.components.price.PriceColor.loaderColor
-import com.miam.kmm_miam_sdk.android.ui.components.price.PriceColor.priceDecimalColor
-import com.miam.kmm_miam_sdk.android.ui.components.price.PriceColor.priceDecimalTotalPriceColor
 import com.miam.kmm_miam_sdk.android.ui.components.price.PriceColor.priceIntegerColor
 import com.miam.kmm_miam_sdk.android.ui.components.price.PriceColor.priceIntegerTotalPriceColor
 import com.miam.kmm_miam_sdk.android.ui.components.price.PriceColor.subtitleColor
@@ -31,54 +33,52 @@ import com.miam.kmm_miam_sdk.android.ui.components.price.PriceStyle.priceEmptySt
 import com.miam.kmm_miam_sdk.android.ui.components.price.PriceText.currency
 import com.miam.kmm_miam_sdk.android.ui.components.price.PriceText.preGuests
 import com.miam.kmm_miam_sdk.android.ui.components.states.ManagementResourceState
+import java.math.RoundingMode
+import java.text.DecimalFormat
 
-class Price(
-    val recipeId: String? = "",
-    private val guestNumber: Int? = -1,
-    val price: Double? = null,
-    val isTotalPrice: Boolean = false,
+@Composable
+fun Price(
+    recipeId: String? = "",
+    guestNumber: Int = -1,
+    price: Double? = null,
+    isTotalPrice: Boolean = false,
 ) {
+    val vmPrice =
+        PricingViewModel()
 
-    private var vmPrice: PricingViewModel = PricingViewModel()
-
-    init {
-        if ((recipeId != "" || recipeId != null) && guestNumber != -1) {
-            vmPrice.setEvent(
-                PricingContract.Event.OnSetRecipe(recipeId!!, guestNumber!!)
-            )
-        } else if (price != null) {
-            vmPrice.setEvent(
-                PricingContract.Event.SetDirectPrice(price)
-            )
-        }
-
+    if (recipeId != "" && guestNumber != -1) {
         vmPrice.setEvent(
-            PricingContract.Event.OnPriceUpdate
+            PricingContract.Event.OnSetRecipe(recipeId!!, guestNumber)
+        )
+    } else if (price != null) {
+        vmPrice.setEvent(
+            PricingContract.Event.SetDirectPrice(price)
         )
     }
 
-    @Composable
-    fun content() {
-        val state by vmPrice.uiState.collectAsState()
-        Box {
-            ManagementResourceState(
-                resourceState = state.price,
-                successView = { price ->
-                    requireNotNull(price)
-                    PriceView(
-                        vmPrice.currentState.integerPart,
-                        vmPrice.currentState.decimalPart,
-                        isTotalPrice
-                    )
-                },
-                emptyView = { EmptyState() },
-                onTryAgain = { /*TODO*/ },
-                onCheckAgain = { /*TODO*/ },
-                loadingView = { PriceShimmer(isTotalPrice) }
-            )
-        }
-    }
+    PriceStateManager(vmPrice, isTotalPrice)
 
+}
+
+@Composable
+fun PriceStateManager(vmPrice: PricingViewModel, isTotalPrice: Boolean) {
+    val state by vmPrice.uiState.collectAsState()
+    Box {
+        ManagementResourceState(
+            resourceState = state.price,
+            successView = { price ->
+                requireNotNull(price)
+                PriceView(
+                    price.price,
+                    isTotalPrice
+                )
+            },
+            emptyView = { EmptyState() },
+            onTryAgain = { /*TODO*/ },
+            onCheckAgain = { /*TODO*/ },
+            loadingView = { PriceShimmer(isTotalPrice) }
+        )
+    }
 }
 
 @Composable
@@ -87,20 +87,18 @@ fun EmptyState() {
 }
 
 @Composable
-fun PriceView(integerPart: String, decimalPart: String, isTotalPrice: Boolean) {
+fun PriceView(price: Double, isTotalPrice: Boolean) {
+    val df = DecimalFormat("#.##")
+    df.roundingMode = RoundingMode.DOWN
+
     Column(
         modifier = mainContainer,
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
         Row(modifier = priceContainer) {
             Text(
-                "${integerPart},",
+                "${df.format(price)}$currency",
                 color = if (isTotalPrice) priceIntegerTotalPriceColor else priceIntegerColor,
-                style = subtitleBold
-            )
-            Text(
-                "${decimalPart}$currency",
-                color = if (isTotalPrice) priceDecimalTotalPriceColor else priceDecimalColor,
                 style = subtitleBold
             )
         }
@@ -172,13 +170,13 @@ fun shimmerPriceItem(brush: Brush, isTotalPrice: Boolean) {
 @Preview
 @Composable
 fun PricePreview() {
-    PriceView("10", "05", false)
+    PriceView(10.0, false)
 }
 
 @Preview
 @Composable
 fun TotalPricePreview() {
-    PriceView("10", "05", true)
+    PriceView(10.0, true)
 }
 
 @Preview