+++ androidSDK/src/main/java/com/miam/kmm_miam_sdk/android/ui/components/price/PriceStyle.kt
@@ -6,11 +6,10 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.unit.dp
 import com.miam.kmm_miam_sdk.android.theme.Dimension.lSpacerHeight
 import com.miam.kmm_miam_sdk.android.theme.Dimension.mRoundedCorner
-import com.miam.kmm_miam_sdk.android.theme.Dimension.mSpacerHeight
 import com.miam.kmm_miam_sdk.android.theme.Dimension.sPadding
-import com.miam.kmm_miam_sdk.android.theme.Dimension.sSpacerHeight
 import com.miam.kmm_miam_sdk.android.theme.Dimension.sSpacerWidth
 
 
@@ -23,12 +22,7 @@ object PriceStyle {
 
     var loaderInteger: Modifier = Modifier
         .padding(vertical = sPadding)
-        .height(mSpacerHeight)
-        .width(sSpacerWidth)
-        .clip(RoundedCornerShape(mRoundedCorner))
-
-    var loaderDecimal: Modifier = Modifier
-        .height(sSpacerHeight)
+        .height(22.dp)
         .width(sSpacerWidth)
         .clip(RoundedCornerShape(mRoundedCorner))